requirements.txt
dulwich
python-dateutil
django
south
django-secure